requirements.txt
# Core API
fastapi
uvicorn[standard]
pydantic
pydantic-settings

# Web Scraping
requests
beautifulsoup4
lxml

# Data Processing
pandas
numpy

# Authentication
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# Database (opcional para futuro)
sqlalchemy

# Monitoring & Logging
python-json-logger

# CORS
python-dotenv

# Dashboard
streamlit
plotly

# Metricas
prometheus-fastapi-instrumentator
